refactor(genart): better naming for background
